requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
astroid==2.11.5
async-timeout==4.0.2
attrs==21.4.0
base58==2.1.1
bitarray==1.2.2
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
cytoolz==0.11.2
dill==0.3.5.1
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
frozenlist==1.3.0
hexbytes==0.2.2
idna==3.3
ipfshttpclient==0.8.0a2
isort==5.10.1
jsonschema==4.4.0
lazy-object-proxy==1.7.1
lru-dict==1.1.7
mccabe==0.7.0
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
netaddr==0.8.0
parsimonious==0.8.1
pathspec==0.9.0
platformdirs==2.5.2
protobuf==3.20.1
pycryptodome==3.14.1
pylint==2.14.0
pyrsistent==0.18.1
requests==2.27.1
rlp==2.0.1
six==1.16.0
tomli==2.0.1
tomlkit==0.11.0
toolz==0.11.2
urllib3==1.26.9
varint==1.0.2
web3==5.29.0
websocket-client==1.3.2
websockets==9.1
wrapt==1.14.1
yarl==1.7.2
